requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0

# HEP data loading
uproot>=5.0.0
awkward>=2.0.0
vector>=1.0.0

# Lorentz-equivariant transformer components (L-GATr)
# NOTE: use 'lgatr' (not 'lorentz-gatr') — different package, correct module name
# pip install lgatr
lgatr>=1.4.0
einops>=0.6.0      # required by lgatr

# Experiment utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.28.0   # used by data download utilities

# Evaluation and visualisation
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebook (for demo.ipynb)
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0
